refactor: moving commands to mdchat.commands package
